% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblefunction.R
\name{geom_scribblefunction}
\alias{geom_scribblefunction}
\title{Create a ggplot layer containing scribbled function paths}
\usage{
geom_scribblefunction(
  mapping = NULL,
  data = NULL,
  stat = "function",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  res = 200
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{Ignored by \code{stat_function()}, do not use.}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{res}{An approximate measure of the "resolution" of the wibbled lines
in the layer. It can be thought of as the "frequency" of the wibble,
much as \code{wibbliness} is the "amplitude". Values around the default of
200 will give a rough, hand-drawn look. Lower numbers are neater and
higher numbers are "shakier"}
}
\value{
A `Layer` ggproto object that can be added to a plot.
}
\description{
Create a ggplot layer containing scribbled function paths
}
\section{Aesthetics}{

\code{geom_scribblefunction()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \strong{\code{y}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{group}
  \item \code{linetype}
  \item \code{linewidth}
  \item \code{wibbliness}
  \item \code{wonkiness}
}


Learn more about setting these aesthetics 
in \code{vignette("ggplot2-specs")}.
}

\examples{
library(ggplot2)

ggplot() +
  geom_scribblefunction(fun = dnorm, colour = "red") +
  xlim(-5, 5)
}
