% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblepolygon.R
\name{geom_scribblepolygon}
\alias{geom_scribblepolygon}
\title{Create a ggplot layer containing scribble-filled polygons}
\usage{
geom_scribblepolygon(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  linejoin = "mitre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by `aes()`. If specified and
`inherit.aes = TRUE` (the default), it is combined with the default mapping
at the top level of the plot. You must supply `mapping` if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If `NULL`, the default, the data is inherited from the plot
   data as specified in the call to `ggplot()`.

   A `data.frame`, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   `fortify()` for which variables will be created.

   A `function` will be called with a single argument,
   the plot data. The return value must be a `data.frame`, and
   will be used as the layer data. A `function` can be created
   from a `formula` (e.g. `~ head(.x, 10)`).}

\item{stat}{The statistical transformation to use on the data for this layer,
either as a `ggproto` `Geom` subclass or as a string naming the stat
stripped of the `stat_` prefix (e.g. "count" rather than `"stat_count"`)}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. "jitter" to use position_jitter), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed to `layer()`}

\item{na.rm}{If FALSE, the default, missing values are removed with a
warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
`NA`, the default, includes if any aesthetics are mapped.
`FALSE` never includes, and `TRUE` always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use `TRUE`.  If `NA`, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. `borders()`.}
}
\description{
Create a ggplot layer containing scribble-filled polygons
}
\examples{
ggplot2::ggplot(data = data.frame(x = c(5, 10, 7.5, 10, 15, 12.5),
                                  y = c(5, 5, 10, 5, 5, 10),
                                  g = rep(1:2, each = 3)),
                mapping = ggplot2::aes(x, y, group = g)) +
 geom_scribblepolygon()

}
