% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblearea.R
\docType{data}
\name{GeomScribblearea}
\alias{GeomScribblearea}
\title{The ggproto object that powers scribble-filled area polygons.}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
