% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblecontour.R
\docType{data}
\name{GeomScribblecontour}
\alias{GeomScribblecontour}
\title{The ggproto object that powers scribbled 2d contour lines}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
