% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblecontour_filled.R
\docType{data}
\name{GeomScribblecontourFilled}
\alias{GeomScribblecontourFilled}
\title{The ggproto object that powers scribble-filled contours}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
