% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblecrossbar.R
\docType{data}
\name{GeomScribblecrossbar}
\alias{GeomScribblecrossbar}
\title{The ggproto object that powers scribbled crossbars}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
