% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribbledensity.R
\docType{data}
\name{GeomScribbledensity}
\alias{GeomScribbledensity}
\title{The ggproto object that powers scribbled kernel density estimates}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
