% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribbledensity2d.R
\docType{data}
\name{GeomScribbledensity2d}
\alias{GeomScribbledensity2d}
\title{The ggproto object that powers scribbled 2d density lines}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
