% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribbledensity2d_filled.R
\docType{data}
\name{GeomScribbledensity2dFilled}
\alias{GeomScribbledensity2dFilled}
\title{The ggproto object that powers scribble-filled density contours}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
