% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribbleerrorbarh.R
\docType{data}
\name{GeomScribbleerrorbarh}
\alias{GeomScribbleerrorbarh}
\title{The ggproto object that powers scribbled horizontal error bars}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
