% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblefunction.R
\docType{data}
\name{GeomScribblefunction}
\alias{GeomScribblefunction}
\title{The ggproto object that powers scribbled function paths}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
