% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblehline.R
\docType{data}
\name{GeomScribblehline}
\alias{GeomScribblehline}
\title{The ggproto object that powers scribbled horizontal reference lines}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
