% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblesegment.R
\docType{data}
\name{GeomScribblelinerange}
\alias{GeomScribblelinerange}
\title{The ggproto object that powers scribbled lineranges}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
