% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblepointrange.R
\docType{data}
\name{GeomScribblepointrange}
\alias{GeomScribblepointrange}
\title{The ggproto object that powers scribble-filled pointranges}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
