% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblepolygon.R
\docType{data}
\name{GeomScribblePolygon}
\alias{GeomScribblePolygon}
\title{The ggproto object that powers scribble-filled shapes.}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
