% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblequantile.R
\docType{data}
\name{GeomScribblequantile}
\alias{GeomScribblequantile}
\title{The ggproto object that powers scribbled quantile lines}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
