% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblerect.R
\docType{data}
\name{GeomScribbleRect}
\alias{GeomScribbleRect}
\title{The ggproto object that powers scribble-filled rectangles.}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
