% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribbleribbon.R
\docType{data}
\name{GeomScribbleribbon}
\alias{GeomScribbleribbon}
\title{The ggproto object that powers scribble-filled ribbons}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
