% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblesmooth.R
\docType{data}
\name{GeomScribblesmooth}
\alias{GeomScribblesmooth}
\title{The ggproto object that powers scribbled conditional means}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
