% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribbleviolin.R
\docType{data}
\name{GeomScribbleviolin}
\alias{GeomScribbleviolin}
\title{The ggproto object that powers scribbled violin plots}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
