% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scribblevline.R
\docType{data}
\name{GeomScribblevline}
\alias{GeomScribblevline}
\title{The ggproto object that powers scribbled vertical reference lines}
\description{
See \link[ggplot2]{ggplot2-ggproto}
}
\keyword{datasets}
