% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_guide.R, R/geom_scribbleabline.R,
%   R/geom_scribblearea.R, R/geom_scribblebar.R, R/geom_scribbleboxplot.R,
%   R/geom_scribblecontour.R, R/geom_scribblecontour_filled.R,
%   R/geom_scribblecrossbar.R, R/geom_scribbledensity.R,
%   R/geom_scribbledensity2d.R, R/geom_scribbledensity2d_filled.R,
%   R/geom_scribbleerrorbar.R, R/geom_scribbleerrorbarh.R,
%   R/geom_scribblefunction.R, R/geom_scribblehex.R, R/geom_scribblehline.R,
%   R/geom_scribbleline.R, R/geom_scribblelinerange.R, R/geom_scribblepath.R,
%   R/geom_scribblepoint.R, R/geom_scribblepointrange.R,
%   R/geom_scribblepolygon.R, R/geom_scribblequantile.R, R/geom_scribblerect.R,
%   R/geom_scribbleribbon.R, R/geom_scribblerug.R, R/geom_scribblesegment.R,
%   R/geom_scribblesmooth.R, R/geom_scribblespoke.R, R/geom_scribblestep.R,
%   R/geom_scribbletile.R, R/geom_scribbleviolin.R, R/ggproto.R
\docType{data}
\name{GuideScribbleaxis}
\alias{GuideScribbleaxis}
\alias{GeomScribbleabline}
\alias{GeomScribblearea}
\alias{GeomScribblebar}
\alias{GeomScribblecol}
\alias{GeomScribbleboxplot}
\alias{GeomScribblecontour}
\alias{GeomScribblecontourFilled}
\alias{GeomScribblecrossbar}
\alias{GeomScribbledensity}
\alias{GeomScribbledensity2d}
\alias{GeomScribbledensity2dFilled}
\alias{GeomScribbleerrorbar}
\alias{GeomScribbleerrorbarh}
\alias{GeomScribblefunction}
\alias{GeomScribblehex}
\alias{GeomScribblehline}
\alias{GeomScribbleline}
\alias{GeomScribblelinerange}
\alias{GeomScribblepath}
\alias{GeomScribblepoint}
\alias{GeomScribblepointrange}
\alias{GeomScribblepolygon}
\alias{GeomScribblequantile}
\alias{GeomScribblerect}
\alias{GeomScribbleribbon}
\alias{GeomScribblerug}
\alias{GeomScribblesegment}
\alias{GeomScribblesmooth}
\alias{GeomScribblespoke}
\alias{GeomScribblestep}
\alias{GeomScribbletile}
\alias{GeomScribbleviolin}
\alias{ggscribble-ggproto}
\title{"ggproto" objects exported from the "ggscribble" package}
\description{
The \code{ggscribble} package defines several \code{ggproto} objects for
creating plots with a hand-drawn appearance. These objects serve as the
foundation for the geoms, scales and guides in this package. The casual user
does not need to know anything about these objects, but they are exported
so that they can be explored and used by ggplot extension developers.
}
\details{
ggproto objects in ggplot2 define the behavior of geoms, stats, and other
components by encapsulating methods and default settings. They allow for
inheritance and method overriding, making it easier to create new geoms
and stats by building upon existing ones. For a detailed description see
 \code{\link[ggplot2]{ggplot2-ggproto}}.
}
\keyword{datasets}
