% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscribble-package.R
\docType{package}
\name{ggscribble-package}
\alias{ggscribble}
\alias{ggscribble-package}
\title{ggscribble: Create Plots With A Hand-Drawn Look}
\description{
A `ggplot2` extension that allows plots to achieve a hand-drawn appearance via custom layers, guides, theme elements and fonts.
}
\author{
\strong{Maintainer}: Allan Cameron \email{Allan.Cameron@nhs.scot}

}
\keyword{internal}
