% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_guide.R
\name{guide_scribbleaxis}
\alias{guide_scribbleaxis}
\title{Title}
\usage{
guide_scribbleaxis(
  title = ggplot2::waiver(),
  theme = NULL,
  check.overlap = FALSE,
  angle = ggplot2::waiver(),
  n.dodge = 1,
  minor.ticks = FALSE,
  cap = "none",
  order = 0,
  position = ggplot2::waiver(),
  wibbliness = 1,
  res = 200
)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link[ggplot2:waiver]{waiver()}}), the name of the scale object or the name
specified in \code{\link[ggplot2:labs]{labs()}} is used for the title.}

\item{theme}{A \code{\link[ggplot2:theme]{theme}} object to style the guide individually or
differently from the plot's theme settings. The \code{theme} argument in the
guide overrides, and is combined with, the plot's theme.}

\item{check.overlap}{silently remove overlapping labels,
(recursively) prioritizing the first, last, and middle labels.}

\item{angle}{Compared to setting the angle in \code{\link[ggplot2:theme]{theme()}} / \code{\link[ggplot2:element_text]{element_text()}},
this also uses some heuristics to automatically pick the \code{hjust} and \code{vjust} that
you probably want. Can be one of the following:
\itemize{
\item \code{NULL} to take the angles and \code{hjust}/\code{vjust} directly from the theme.
\item \code{waiver()} to allow reasonable defaults in special cases.
\item A number representing the text angle in degrees.
}}

\item{n.dodge}{The number of rows (for vertical axes) or columns (for
horizontal axes) that should be used to render the labels. This is
useful for displaying labels that would otherwise overlap.}

\item{minor.ticks}{Whether to draw the minor ticks (\code{TRUE}) or not draw
minor ticks (\code{FALSE}, default).}

\item{cap}{A \code{character} to cut the axis line back to the last breaks. Can
be \code{"none"} (default) to draw the axis line along the whole panel, or
\code{"upper"} and \code{"lower"} to draw the axis to the upper or lower break, or
\code{"both"} to only draw the line in between the most extreme breaks. \code{TRUE}
and \code{FALSE} are shorthand for \code{"both"} and \code{"none"} respectively.}

\item{order}{A positive \code{integer} of length 1 that specifies the order of
this guide among multiple guides. This controls in which order guides are
merged if there are multiple guides for the same position. If 0 (default),
the order is determined by a secret algorithm.}

\item{position}{Where this guide should be drawn: one of top, bottom,
left, or right.}

\item{wibbliness}{How wibbly the axis lines should be (0 = straight)}

\item{res}{A measure of the resolution of the line's wibbliness.
Default is 200. Higher numbers give a higher frequency wibble}
}
\value{
A guide object that draws wibbly axis lines and ticks
}
\description{
Title
}
\examples{
ggplot(mtcars, aes(wt, mpg)) +
  geom_scribblepoint() +
  ggplot2::guides(x = guide_scribbleaxis(wibbliness = 0.5),
                  y = guide_scribbleaxis(wibbliness = 0.5)) +
  theme_classic()
}
