% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_scribblecolour.R
\name{scale_scribblecolour_gradient}
\alias{scale_scribblecolour_gradient}
\alias{scale_scribblecolour_continuous}
\alias{scale_scribblecolour_identity}
\alias{scale_scribblecolour_manual}
\alias{scale_scribblecolour_discrete}
\title{Scribble colour scales}
\usage{
scale_scribblecolour_gradient(
  name = waiver(),
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "scribblecolour"
)

scale_scribblecolour_continuous(
  name = waiver(),
  ...,
  low = "#132B43",
  high = "#56B1F7",
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "scribblecolour"
)

scale_scribblecolour_identity(
  name = waiver(),
  ...,
  guide = "none",
  aesthetics = "scribblecolour"
)

scale_scribblecolour_manual(
  ...,
  values,
  aesthetics = "scribblecolour",
  breaks = waiver(),
  na.value = "grey50"
)

scale_scribblecolour_discrete(
  name = waiver(),
  ...,
  h = c(0, 360) + 15,
  c = 100,
  l = 65,
  h.start = 0,
  direction = 1,
  na.value = "grey50",
  aesthetics = "scribblecolour"
)
}
\arguments{
\item{...}{Other arguments passed on to `continuous_scale()`, `binned_scale`,
or `discrete_scale()` as appropriate, to control name, limits,
breaks, labels and so forth.}

\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{A function used to create a guide or its name. See
[guides()] for more information.}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with breaks if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that
don't match will be given na.value.}

\item{breaks}{One of:
- `NULL` for no breaks
- `waiver()` for the default breaks computed by the
  [transformation object][scales::trans_new()]
- A numeric vector of positions
- A function that takes the limits as input and returns breaks
  as output (e.g., a function returned by [scales::extended_breaks()]).
  Also accepts rlang [lambda][rlang::as_function()] function notation.}

\item{range}{Output range of angle in degrees}
}
\value{
A `Scale` ggproto object that can be added to a plot.
}
\description{
These functions allow numeric or categorical variables to be passed to
the aesthetic mapping of a plot layer to change the colour of the scribble
fill lines. These scales are directly analogous to the scale functions
in ggplot.
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(Species, scribblecolour = Species,
                                   colour = Species)) +
  geom_scribblebar() +
  scale_scribblecolour_manual(values = c("green4", "navy", "red3")) +
  ggplot2::scale_colour_manual(values = c("green4", "navy", "red3"))
}
