% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_scribbledensity.R
\name{scale_scribbledensity}
\alias{scale_scribbledensity}
\alias{scale_scribbledensity_continuous}
\alias{scale_scribbledensity_identity}
\alias{scale_scribbledensity_manual}
\alias{scale_scribbledensity_ordinal}
\alias{scale_scribbledensity_discrete}
\title{Scribble density scales}
\usage{
scale_scribbledensity(
  name = ggplot2::waiver(),
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = NULL,
  range = c(50, 400),
  guide = "legend"
)

scale_scribbledensity_continuous(
  name = ggplot2::waiver(),
  breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = NULL,
  range = c(50, 400),
  guide = "legend"
)

scale_scribbledensity_identity(name = ggplot2::waiver(), ..., guide = "none")

scale_scribbledensity_manual(
  ...,
  values,
  breaks = ggplot2::waiver(),
  na.value = NA
)

scale_scribbledensity_ordinal(
  name = ggplot2::waiver(),
  ...,
  range = c(50, 400)
)

scale_scribbledensity_discrete(...)
}
\arguments{
\item{breaks}{One of:
- `NULL` for no breaks
- `waiver()` for the default breaks computed by the
  [transformation object][scales::trans_new()]
- A numeric vector of positions
- A function that takes the limits as input and returns breaks
  as output (e.g., a function returned by [scales::extended_breaks()]).
  Also accepts rlang [lambda][rlang::as_function()] function notation.}

\item{range}{Output range of density (higher values are more closely packed)}

\item{guide}{A function used to create a guide or its name. See
[guides()] for more information.}

\item{...}{Other arguments passed on to `continuous_scale()`, `binned_scale`,
or `discrete_scale() as appropriate, to control name, limits,
breaks, labels and so forth.}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with breaks if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that
don't match will be given na.value.}

\item{na.value}{Missing values will be replaced with this value.}
}
\value{
A `Scale` ggproto object that can be added to a plot.
}
\description{
These functions allow numeric or categorical variables to be passed to
the aesthetic mapping of a plot layer to change the density of the scribble
fill lines. These scales are directly analogous to the scale functions
in ggplot.
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(Species, scribblecolour = Species,
                                   colour = Species,
                                   scribbledensity = Species)) +
 geom_scribblebar() +
 scale_scribbledensity_manual(values = c(45, 135, 0))
}
