% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_scribblewidth.R
\name{scale_scribblewidth}
\alias{scale_scribblewidth}
\alias{scale_scribblewidth_continuous}
\alias{scale_scribblewidth_identity}
\alias{scale_scribblewidth_manual}
\alias{scale_scribblewidth_ordinal}
\alias{scale_scribblewidth_discrete}
\title{Scribble linewidth scales}
\usage{
scale_scribblewidth(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  guide = "legend"
)

scale_scribblewidth_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  guide = "legend"
)

scale_scribblewidth_identity(name = waiver(), ..., guide = "none")

scale_scribblewidth_manual(..., values, breaks = waiver(), na.value = NA)

scale_scribblewidth_ordinal(name = waiver(), ..., range = c(1, 6))

scale_scribblewidth_discrete(...)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
\link[scales:new_transform]{transformation object}
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output (e.g., a function returned by \code{\link[scales:breaks_extended]{scales::extended_breaks()}}).
Also accepts rlang \link[rlang:as_function]{lambda} function notation.
}}

\item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{limits}{One of:
\itemize{
\item \code{NULL} to use the default scale range
\item A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum
\item A function that accepts the existing (automatic) limits and returns
new limits. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
Note that setting limits on positional scales will \strong{remove} data outside of the limits.
If the purpose is to zoom, use the limit argument in the coordinate system
(see \code{\link[ggplot2:coord_cartesian]{coord_cartesian()}}).
}}

\item{range}{The output range of the scribbled linewidth}

\item{guide}{A function used to create a guide or its name. See
\code{\link[ggplot2:guides]{guides()}} for more information.}

\item{...}{Other arguments to be passed to the relevant scale generator}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the scale,
or with breaks if provided. If this is a named vector, then the values will
be matched based on the names instead. Data values that don't match will be
given na.value}

\item{na.value}{Missing values will be replaced with this value.}
}
\value{
A `Scale` ggproto object that can be added to a plot.
}
\description{
These functions allow numeric or categorical variables to be passed to
the aesthetic mapping of a plot layer to change the width of the scribble
fill lines. These scales are directly analogous to the scale functions
in ggplot.
}
\examples{
library(ggplot2)

ggplot(iris, aes(Species, scribblecolour = Species,
       colour = Species, scribblewidth = Species)) +
 geom_scribblebar() +
 scale_scribblewidth_manual(values = c(1, 3, 5))
}
